Finished backend work. Some of the outputs might want to round the frequencies to the nearest 100th

+++ neureset/Headset.h
@@ -40,7 +40,7 @@ private:
 
     int currentNodeIndex;
     float elapsedTime;
-    int treatmentFreq;
+    float treatmentFreq;
 
 private slots:
     void applyTreatmentToCurrNode();
